refactor: 컴포넌트 구조 분리 및 코드 정리
